Split ResourceList stories into 2 more manageable files
